refactor: extract board projects module
